Outline of server component.
The end product of harmony is two components: a lovely
frontend wherein many people collaborate on a peculiar music
and a server component that concerns itself only with not
failing and keeping track of all entities in the simulation.

This is just a skeleton, one that will not compile. Soon it
will be a real boy.