Begin protocol definition, reorganized harmony code.
Of most interest are two new files: harmonyd.py and tuner.py
in client/test/. These, respectively, implement the harmony
protocol prototype and test the prototype. As of this commit,
the prototype _seems_ to function in a somewhat reasonable
manner, but the validity of its broadcast functionality is
unexplored. Tomorrow, possibly, will have the prototype
finished out.

The harmony codebase was modified to be more idiomatic: tabs
were converted to spaces, getter/setters were removed and
indentation made into four spaces.

Updated org documents to include Google Analytics, just
because.