Backed off on custom protocol: now base64 encoded json.
While not permanent, the current state of affairs means the
Harmony prototype is quicker off the floor, easier to tinker
with and more likely to be hackable by one for a few
individuals. The actual protocol/server will need to be
developed in a branch, porting Cameron's changes as he
progresses.